\name{DistributionOptimization-package}
\alias{DistributionOptimization-package}
\docType{package}
\title{Distribution Optimization}

\description{Distribution Optimization fits  gaussian mixture models on to one dimensional samples by minimizing the Chi Squared Error by evolutional optimization. It is an alternative to likelihood maximizers like expection maximization. Through the included "Overlapping" Methods, single gaussians can be forced to be separated, achieving various signifcant models to choose from. The evolutionary part is done through the "GA" Package. The Gaussian Mixture Logic is based on the "AdaptGauss" Package.
}

\references{
Luca Scrucca (2013). GA: A Package for Genetic Algorithms in R. Journal of Statistical Software, 53(4), 1-37. URL http://www.jstatsoft.org/v53/i04/
}

\author{Florian Lerch, Jorn Lotsch, Alfred Ultsch}

\keyword{EM}
\keyword{GMM}
\keyword{gaussian mixture model}
\keyword{pde}
\keyword{pareto density estimation}
